move requireOut to test utils
